Fix re-parsing with metadata
Previously, the tool could reformat fields, and optionally reformat them
to use metadata={...} dicts. But if you tried to re-parse the fixed script,
it would fail to parse the metadata dictionary. This was because the
`repr_ast` function converted the whole dictionary to a
representation of the dict. Moreover, if metadata was already present in
the field kwargs, it wasn't removed before being appended to.

The new code, recursively unwraps the kwargs dict incase (in the case of
the metadata dict) any of the values is a dict, and any of those values is
a dict, etc.