Default to not sorting, everywhere
Previously, the program-wide default was to not sort kwargs and dictionary
keys. However, each individual formatting function defined the default
sort function to *do* sorting. This meant the unit tests were slightly
different to running the default program.

This commit corrects the issue, by defaulting to a no-op, no-sort function.