Handle comments
Previously, any comments within a field definition would cause the tool to
fail, as the field was condensed down to a single line. To fix this, the
field definition is now kept with it's existing newline spacing until it
is parsed by AST (AST can handle multi-line expressions, unsurprisingly).

However, this then results in comments being stripped out as AST has no
view of comments (they are stripped out by the parser). This means we have
to re-insert the comments ourselves. This is achieved by doing a diff
between the original lines and the new lines, and identifying removed lines
(prefixed by "- ") whose first non-whitespace character is
indicating a comment, and including those lines in the output.

Note that this is not a complete solution, as it does not handle comments
which follow "real" code on the same line.