Validate syntax before writing
The tool is not perfect, and can make incorrect alterations to its target
files. In order to prevent bad files being created, this commit has the CLI
validate the reformatted lines using AST.

If invalid, the tool prints an error message including the SyntaxError
from the AST parse and the failed line with context lines in order to help
better debugging, given that that the file is no longer written.