Fix kwargs containing a dictionary
We didn't do this properly before, because we passed the dictionary to
`maybe_wrap_line` which was expecting a string. The value was just cast to
string naively which meant that any string values were single quoted, not
double. This also exposed that we weren't stripping of quotes from the
values when the dictionary was created.