Handle double quotes within strings
We always quote strings using double quotes. Sometimes strings can contain
double quotes, which means they need to be escaped in order to get valid
syntax.

Note that *all* the `second_bit` arguments that `maybe_wrap_lines` receives
are strings as far as the runtime of the tool is concerned. However, by
this time any to-be-written string values have been quoted with double
quotes. We use those to determine the "type" of the value and then strip
them off before replacing the double quotes, because we don't want the
outer quotes to be escape.