Fix max-line-width handling
Basically every "format_" function requires an extra level on indentation
to the last. Sometimes this was done by the caller, and sometimes by the
called function. It makes most sense for the caller to always decrease the
remaining line length, by one level.

This also highlights that the previous wrapping was wrong in some cases.
The fixes are reflected in the altered unit-tests.